Inspect methods added to Pet and Shelter.
Why:

When creating a new object in the console an overwhelming amount of xml information is returned. The text contains useful information but it is difficult to decipher it from the large quantity of text generated.

How:

This commit adds an inspect method for both the Pet and Shelter classes. The inspect method returns a string containing a few xml attributes when the object is called. For the shelter object the string contains the shelter_id, name and state. For the pet object the string contains the pet_id, name, animal, mix and sex.